requirements.txt
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
flask-restx==0.3.0
gunicorn
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2==2.8.6
PyMySQL==1.0.2
pyrsistent==0.17.3
pytz==2021.1
six==1.15.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
